violet/script/requirements.txt
google-api-python-client==2.130.0
google-cloud-bigquery==3.23.1